refactor: add samples folder
