+++ src/components/samples/Hero.tsx
