+++ src/styles/components/samples/Description.module.scss
