+++ src/styles/components/samples/Hero.module.scss
