+++ src/styles/components/samples/Hotlinks.module.scss
